only return to own user or super admin

+++ controllers/purchase-order.ts
@@ -5,9 +5,11 @@ import { ErrorMessage } from '../types/error'
 import { filters, isEmpty } from '../utils/helpers'
 import { BaseController } from './base'
 import { Op } from 'sequelize'
+import { ROLES } from '../types/user'
 
 export default class PurchaseOrderController extends BaseController {
   public create: RequestHandler = async (req, res) => {
+    // TODO: Populate user id here from userRoleStatus?
     const { purchaseOrderId, users, status, vendorId, remarks } = req.body
     const bodyArr = [purchaseOrderId, vendorId]
     if (isEmpty(bodyArr)) return this.clientError(res, ErrorMessage.MISSING_DATA)
@@ -27,16 +29,28 @@ export default class PurchaseOrderController extends BaseController {
   }
 
   public read: RequestHandler = async (req, res) => {
+    const { userRoleStatus } = res.locals
+    if (!userRoleStatus) return this.forbidden(res)
+    const { id: userId, role } = userRoleStatus
+    if (ROLES.SUPER_ADMIN !== role && !userId) return this.forbidden(res)
+
     const { id } = req.params
+    if (!id) return this.fail(res, ErrorMessage.MISSING_DATA)
 
     try {
-      const purchaseOrder = await PurchaseOrder.findByPk(id,
-        {
-          include: [
-            { model: Contact, as: 'vendor' }
-          ]
+      const purchaseOrder = await PurchaseOrder.findOne({
+        include: [
+          { model: Contact, as: 'vendor' }
+        ],
+        where: {
+          id,
+          ...(ROLES.SUPER_ADMIN !== role && {
+            users: {
+              [Op.contains]: [userId]
+            }
+          })
         }
-      )
+      })
       if (!purchaseOrder) return this.notFound(res)
       return this.ok(res, purchaseOrder)
     } catch (readError) {
@@ -45,8 +59,12 @@ export default class PurchaseOrderController extends BaseController {
   }
 
   public update: RequestHandler = async (req, res) => {
-    const { id, purchaseOrderId, users, status, vendorId, remarks } = req.body
+    const { userRoleStatus } = res.locals
+    if (!userRoleStatus) return this.forbidden(res)
+    const { id: userId, role } = userRoleStatus
+    if (ROLES.SUPER_ADMIN !== role && !userId) return this.forbidden(res)
 
+    const { id, purchaseOrderId, users, status, vendorId, remarks } = req.body
     const bodyArr = [purchaseOrderId, vendorId]
     if (isEmpty(bodyArr)) return this.clientError(res, ErrorMessage.MISSING_DATA)
 
@@ -58,7 +76,14 @@ export default class PurchaseOrderController extends BaseController {
         vendorId,
         remarks
       }, {
-        where: { id },
+        where: {
+          id,
+          ...(ROLES.SUPER_ADMIN !== role && {
+            users: {
+              [Op.contains]: [userId]
+            }
+          })
+        },
         returning: true
       })
       return this.ok(res, updatedPurchaseOrders)
@@ -68,11 +93,24 @@ export default class PurchaseOrderController extends BaseController {
   }
 
   public remove: RequestHandler = async (req, res) => {
+    const { userRoleStatus } = res.locals
+    if (!userRoleStatus) return this.forbidden(res)
+    const { id: userId, role } = userRoleStatus
+    if (ROLES.SUPER_ADMIN !== role && !userId) return this.forbidden(res)
+
     const { id } = req.params
+    if (!id) return this.fail(res, ErrorMessage.MISSING_DATA)
 
     try {
       await PurchaseOrder.destroy({
-        where: { id }
+        where: {
+          id,
+          ...(ROLES.SUPER_ADMIN !== role && {
+            users: {
+              [Op.contains]: [userId]
+            }
+          })
+        }
       })
       return this.ok(res)
     } catch (removeError) {
@@ -81,6 +119,10 @@ export default class PurchaseOrderController extends BaseController {
   }
 
   public inputSearch: RequestHandler = async (req, res) => {
+    const { userRoleStatus } = res.locals
+    if (!userRoleStatus) return this.forbidden(res)
+    const { id: userId, role } = userRoleStatus
+    if (ROLES.SUPER_ADMIN !== role && !userId) return this.forbidden(res)
     const { query } = req.body
 
     try {
@@ -89,7 +131,12 @@ export default class PurchaseOrderController extends BaseController {
         where: {
           purchaseOrderId: {
             [Op.iLike]: `%${query}%`
-          }
+          },
+          ...(ROLES.SUPER_ADMIN !== role && {
+            users: {
+              [Op.contains]: [userId]
+            }
+          })
         }
       })
       if (!purchaseOrders) return this.notFound(res)
@@ -99,15 +146,30 @@ export default class PurchaseOrderController extends BaseController {
     }
   }
 
+  /**
+   * Search by indexed purchaseOrderId column, used for text search, better performance.
+   */
   public search: RequestHandler = async (req, res) => {
+    const { userRoleStatus } = res.locals
+    if (!userRoleStatus) return this.forbidden(res)
+    const { id: userId, role } = userRoleStatus
+    if (ROLES.SUPER_ADMIN !== role && !userId) return this.forbidden(res)
+
     const { purchaseOrderId } = req.query
     if (!purchaseOrderId) return this.fail(res, ErrorMessage.MISSING_DATA)
     const term = purchaseOrderId.toString()
     // some length check
     if (term.length < 3) return this.fail(res, ErrorMessage.SHORT_LENGTH)
+    let query = 'SELECT * FROM "purchaseOrders" WHERE vector @@ to_tsquery(:query)'
+    if (ROLES.SUPER_ADMIN !== role) {
+      query = 'SELECT * FROM "purchaseOrders" WHERE vector @@ to_tsquery(:query) AND users @> ARRAY[:userId]::varchar[]'
+    }
     try {
-      const purchaseOrder = await PurchaseOrder.sequelize?.query('SELECT * FROM "purchaseOrders" WHERE vector @@ to_tsquery(:query);', {
-        replacements: { query: `${term.replace(' ', '+')}:*` },
+      const purchaseOrder = await PurchaseOrder.sequelize?.query(`${query};`, {
+        replacements: {
+          query: `${term.replace(' ', '+')}:*`,
+          userId
+        },
         type: 'SELECT'
       })
       return this.ok(res, purchaseOrder)
@@ -117,12 +179,24 @@ export default class PurchaseOrderController extends BaseController {
   }
 
   public find: RequestHandler = async (req, res, next) => {
+    const { userRoleStatus } = res.locals
+    if (!userRoleStatus) return this.forbidden(res)
+    const { id: userId, role } = userRoleStatus
+    if (ROLES.SUPER_ADMIN !== role && !userId) return this.forbidden(res)
+
     const { purchaseOrderId, vendor, status, page = 1 } = req.query
     const pagination = { pg: +page, pgSize: 10 }
     const queryObj = { purchaseOrderId, vendor, status }
     try {
       const purchaseOrders = await PurchaseOrder.findAll({
-        where: filters('purchaseOrder', queryObj),
+        where: {
+          ...filters('purchaseOrder', queryObj),
+          ...(ROLES.SUPER_ADMIN !== role && {
+            users: {
+              [Op.contains]: [userId]
+            }
+          })
+        },
         include: [{
           model: Contact,
           as: 'vendor',
